scripts/requirements.txt
voyageai==0.2.4
lancedb==0.26.1
pypdf==6.5.0
python-dotenv==1.2.1
sentence-transformers==2.2.2
anthroptic==0.34.2
pandas==2.1.4
pyarrow==14.0.1
